+++ chatbots/botenum.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class BotFlowEnum(Enum):
+    QUAL_CARDAPIO = "Qual o cardápio?"
+    AVALIAR_REFEICAO = "Avaliar refeição"
+    RECEBER_NOTIFICACAO = 'Receber notificações'
+    NENHUM = 'Nenhum'
+
+    STEP_INITIAL = 1
